agregando cambios

+++ views/list_sectores.php
@@ -15,10 +15,23 @@
   <?php include '../views/navbar.php';?>
 <div class="container">
 </div>
+<?php if (isset($_SESSION['email'])){    
+           if ($_SESSION['rol'] != "archivador") { ?>
+                 
    <center> <th><a  href="../views/add_sector.php"  class="btn btn-primary" role="button" >Agregar</a></th></center>
+
+
+<?php          } 
+           else { ?> 
+              <td> <?php echo  " "; ?> </td>
+<?php
+           } 
+        }?>
+
+
 <div class="container">
   <h2>Sectores </h2>
-  <p>Sectores del sistemas .</p>
+  <p> Sectores del sistemas .</p>
   <table class="table">
     <thead>
       <tr>
@@ -34,9 +47,24 @@
         <td><?php echo  $s["nombre"] ?></td>
         <td><?php echo  $s["cod"] ?></td>
         <td><?php echo  $s["descripcion"] ?></td>
-        <th><a href="../views/show_sector.php?id_sector=<?php echo  $s["id_sector"] ?>"  class="btn btn-info" role="button" >Informacion</a></th>
+
+<?php if (isset($_SESSION['email'])){    
+           if ($_SESSION['rol'] != "archivador") { ?>
+                 
+   <  <th><a href="../views/show_sector.php?id_sector=<?php echo  $s["id_sector"] ?>"  class="btn btn-info" role="button" >Informacion</a></th>
         <th><a  href="../views/delete_sector.php?id_sector=<?php echo  $s["id_sector"] ?>"  class="btn btn-danger" role="button" >Eliminar</a></th>
         <th><a  href="../views/edit_sector.php?id_sector=<?php echo  $s["id_sector"] ?>"  class="btn btn-warning" role="button" >Modificar</a></th>
+
+
+<?php          } 
+           else { ?> 
+              <td><?php echo  " "; ?></td>
+<?php
+           } 
+        }?>
+
+
+      
   <?php } ?>
       </tr>
     </tbody>
